construction site model update and building model update

+++ src/main/java/pl/grizwold/ogame/modules/buildings/domain/Building.java
@@ -2,9 +2,11 @@ package pl.grizwold.ogame.modules.buildings.domain;
 
 import lombok.Value;
 
+import java.math.BigInteger;
+
 @Value
 public class Building {
-    private int level;
+    private BigInteger level;
     private BuildingType type;
     private String owner;
     private String planetId;
